+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.0.0 2018-02-04
+
+  * Upgraded promise-duplex@3.0.0, promise-readable@3.1.1,
+    promise-writable@3.1.0
+  * Support `import PromisePiping from 'promise-piping'` syntax.
+
 ## v0.5.0 2018-01-22
 
   * Better handling for errors in nested pipings (when stream for piping is